reorganize solution
