+++ client/src/components/InfoTile.tsx
