+++ client/src/routes/_authenticated/achats/mes-encheres.tsx
@@ -3,7 +3,7 @@ import {
   listQualitiesOptions,
 } from '@/api/generated/@tanstack/react-query.gen'
 import { BreadcrumbSection } from '@/components/BreadcrumbSection'
-import AuctionMarketplace from '@/components/auctions/AuctionMarket'
+import AuctionMarketplace from '@/components/auctions/AuctionMarket/AuctionMarket'
 import { useAuthUser } from '@/store/userStore'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
